+++ RELEASES.md
@@ -63,6 +63,7 @@ Upcoming
 - Remove unused and broken Larva jsp's
 - Add scenariosroot.default property
 - Add support to Larva for Maven based Eclipse projects
+- Sort available scenarios root directories before unavailable ones